10. Great leaders are willing to sacrifice the numbers to save the people. Poor leaders sacrifice the people to save the numbers

+++ 0x05-pointers_arrays_strings/100-atoi.c
@@ -13,7 +13,7 @@
  */
 int _atoi(char *s)
 {
-	int i, sign, num;
+	long i, sign, num;
 
 	i = 0;
 	sign = 1;
